+++ src/layout/styles/_layout.scss
@@ -15,6 +15,12 @@ body {
     height: 100%;
   }
 
+  header {
+    > button.close {
+      margin: 0 3px;
+    }
+  }
+
   @import './tabbing';
   @import './split';
