+++ src/layout/styles/_tabbing.scss
@@ -21,16 +21,16 @@ header[data-layout=tabbing] {
   flex: 0 0 auto;
   border-bottom: 1px solid $phovea_layout_active_tab_header;
 
-  > :first-child {
+  > :nth-child(2) {
     display: inline-block;
     margin: 0 2px;
     padding: 3px 1em;
   }
 
-  > :nth-child(n+2) {
+  > header {
     display: inline-block;
     margin: 0 2px;
-    padding: 3px 1em;
+    padding: 3px 3px 3px 1em;
     border-radius: 5px 5px 0 0;
     background-color: $phovea_layout_tab_header;
     border-top: 1px solid $phovea_layout_active_tab_header;
@@ -38,6 +38,11 @@ header[data-layout=tabbing] {
     border-right: 1px solid $phovea_layout_active_tab_header;
     transition: background-color ease 0.5s;
     cursor: pointer;
+
+    > button.close {
+      margin-left: 6px;
+      font-size: 140%;
+    }
 
     &.active {
       cursor: not-allowed;
